Do eight digits and put sum back to 64-bits

+++ Specialization2Digits.cc
@@ -1,5 +1,5 @@
 template<>
-inline void innerBody<2>(std::ostream& stream, uint8_t sum, u64 product, u64 index) noexcept {
+inline void innerBody<2>(std::ostream& stream, u64 sum, u64 product, u64 index) noexcept {
 innerBody<0>(stream, sum+ 4 + 2, product* 4 * 2, index + 24);
 innerBody<0>(stream, sum+ 8 + 2, product* 8 * 2, index + 28);
 innerBody<0>(stream, sum+ 2 + 3, product* 2 * 3, index + 32);
